backend/requirements.txt
asgiref==3.7.2
boto3==1.28.37
botocore==1.31.37
Django==4.2.4
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
environs==9.5.0
geographiclib==2.0
geopy==2.4.0
jmespath==1.0.1
marshmallow==3.20.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
